raisesgroup followups (#13279)
* raisesgroup followups
* renames src/_pytest/raises_group.py to src/_pytest/raises.py
* moves pytest.raises from src/_pytest/python_api.py to src/_pytest/raises.py
* adds several newsfragments that should've been bundled with #13192
* add more detailed error message if you try to do RaisesGroup((ValueError, TypeError))
* mess around with ValueError vs TypeError on invalid expected exception
* revert change in behaviour if raises has a type mismatch
* add check example to raises, fix test after behaviour revert
* made args to AbstractMatcher, RaisesExc and RaisesGroup pos/kw-only

---------
